Write back out to temp file for the secondary instance
After writing out the args to the file, the secondary instance now closes
the file, waits for the primary instance to open it, then polls it until
it can open it again. Once it can, it opens it, reads out the data written
to it by the primary instance, and prints it out

The intention is for the primary instance to write out the appropriate
response to whatever args were passed to the secondary instance, e.g. all
of the register values for "-r" or "--registers".